+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- DocumentSearch.ingest now raises IngestExecutionError when any errors are encountered during ingestion. 
+
 ## 0.12.0 (2025-03-25)
 
 ### Changed